cloudrun/pyproject.toml
[project]
name = "cloudrun"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "google-auth>=2.49.1",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "locust>=2.43.3",
    "pyclean>=3.5.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "respx>=0.22.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
